imgtestlib: add runcmd() variant with capture=False
Add a second runcmd() function, runcmd_nc(), that does not capture the
output.  This is used when we don't care about parsing the output and
would prefer to have it printed in real time during the execution of
cmd.

It is useful to have this as a reusable function to handle the extra_env
and the output handling a bit differently from just check=True.

Use the new command in the few places were we had a separate
implementation or called sp.run() directly.