imgtestlib: add gen_manifests() helper function
Add a new helper/wrapper function for cmd/gen-manifests.
It takes as arguments all the flags that vary between calls and sets the
common arg values so that they are only ever defined in one place.
The common functionality that is abstracted away now is the number of
workers, the cache directory, and setting the OSBUILD_TESTING_RNG_SEED
environment variable.