manifest: set pipeline->manifest ref in addPipeline()
When adding a Pipeline to a Manifest, add the reference that backlinks
from the Pipeline to the Manifest at the same time.

Previously, this was set explicitly in the NewBase() function and a
check was made that the manifest argument was the same as the manifest
for the build pipeline argument (if a build pipeline was provided).  The
manifest.NewBase() function is the constructor for the base pipeline
implementation that every other pipeline embeds.

With this change, a pipeline does not get a reference to a manifest
until it is added to one in the Manifest.addPipeline() method.  Checks
are added to ensure that a pipeline is not added to more than one
manifest (it should have no manifest reference before being added) and
that the build pipeline (if it has one) has a reference to the same
manifest.

The NewBase() function is simplified so that no manifest argument is
needed.  This is the first step in simplifying pipeline constructors.