manifest: call Manifest.addPipeline() in Build.addDependent()
Pipeline constructors that require a build pipeline all add themselves
to the build pipeline's dependents using the Build.addDependent()
method.

All our pipelines (except Build) must define a build pipeline (we always
call Build.addDependent() in their constructors) and the build pipeline
must always have a manifest associated with it if it will have
dependencies.  So now when we create a new pipeline, for example
NewOS(), we don't need to specify the manifest that the pipeline will be
added to because the pipeline constructor *requires* a build pipeline to
depend on and it *must* share a manifest reference with that pipeline
(i.e. the build and OS pipelines must belong to the same manifest if one
depends on the other).