distro: interface that lists supported and required blueprint options
Add a new interface that implements two methods:
- SupportedBlueprintOptions(): Returns a list of strings that indicate
  blueprint options supported by an image type.
- RequiredBlueprintOptions(): Returns a list of strings that indicate
  the blueprint options required by an image type.

The interface is also part of the ImageType interface.  This might
change in the future, but for now we will need it for cases where
ImageType instances need to be validated.

The two fields are added to the ImageTypeYAML loader to support
configuring these for each image type.