osbuild: simplify containers_input.go
While writing tests to ensure the that the `ContainerDeployInputs`
generates the correct json I noticed that this was more involved
than I had hoped and it seems there are some indirections in the
code that may not be necessary. I removed them and IMHHO the code
is now a bit more direct and easier to read.