reporegistry: export ReposByImageTypeName instead of ReposByImageType
Let's export ReposByImageTypeName(), which accepts distro, arch and
image type names as strings, instead of just `distro.ImageType`. This
has the advantage, that the caller may actually use a distro name alias
to get repositories, instead of the actual distro name that the image
type is associated with.

The ReposByImageType() is removed in favor of ReposByImageTypeName().

Signed-off-by: Tomáš Hozza <thozza@redhat.com>