LoadAllRepositories: don't fail on filenames which can't be parsed
Previously, the `LoadAllRepositories()` function used repo file filename
as the distro name, without inspecting it. Now that we parse it, the
failure to parse the filename as the distro ID string was considered an
error. This may cause osbuild-composer fail to start due to failure
while loading repositories.

Let's use the repo file filename as is as the distro name, if parsing it
fails.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>