test: run sudo with -E when building images
When building in CI, we need to preserve the environment so that the
command 'sudo ./bin/build' inherits the rng seed.  Without it, we always
end up building images with a random seed which means that our cache is
updated with metadata that will never match a subsequent manifest check.