test: sync only necessary test info when filtering
When filtering builds for a CI pipeline, we download all the info.json
files from the S3 bucket where we keep our test build cache.  Even
though this isn't a lot of data, the number of files can grow fast which
makes the operation take longer.  Since we only need the info for a
specific distribution and architecture combination, there's no need to
download all of the files every time.

Let's download only the `{distro}/{arch}/` subdirectory that contains
the info.json files that we need.