osbuild: simplify code by removing `osbuild.Mounts type`
The `osbuild.Mounts` type is just an alias to `[]osbuild.Mount`
and it feels like it does not add clarity. From the type it's
not clear if it's a slice or a map (like `osbuild.Devices`). While
it is nice to abstract things away one needs to know it anyway when
constructing the type so this commit removes the type and just
exposes/uses `[]osbuild.Mount` directly.

Another benefit of this is that the pointers used are now clearer
(IMHO) - i.e. `[]Mount` is a slice so no need to indirect it further
via `*Mounts` in the function signatures.