image,manifest: allow using the qcow2 pipeline without a build pipeline
This commit allows to use the qcow2 pipeline outside of a build
pipeline. The use-case is the `bootc-image-builder` that will use
arbitrary containers as buildroots and we cannot assume that they
have `qemu-util` installed. And because `bootc-image-builder` is
already a container we can install `qemu-util` there and be
confident that it will work.