osbuild,manifest: simplify `osbuild.Devices`
The osbuild.Devices type is just an alias to map[string]osbuild.Device
and it feels like it might be easier to expose the map[string]Device
directly. From the type it's not clear if it's a map and when using
the type it needs to be used like map it seems easiest to expose it
directly.

Another benefit of this is that the pointers used are now clearer (IMHO)
i.e. no indirections required for maps.