test: return and print stdout from aws s3 sync
Return the output from the aws s3 sync command and print it only when
running the filter_builds() function, which syncs a limited set of build
info.

When running from check-build-coverage, don't print the output since
that will download all files and generates too much output for the job
log (and isn't very useful).  Also, fail the job if the sync failed
since it doesn't make sense to run anything after it and a failure in
that job could be a sign that something's wrong.

In the filter_builds() function, don't fail if the sync failed, just let
it run and generate a build config for all the image configurations.