disk: add support for btrfs partitioning mode
TODO: needs tests

The conversion code from raw partitioning is basically the same as the one
for LVM. The main difference is that LVM just embeds an already defined
filesystem, but btrfs-ication needs to do some guessing because the code
actually creates a new partition: specifically, I just hardcoded the
compression to zstd:1. This is not ideal, but it's the Fedora's default,
so it's imho fine as the first iteration.