image: unify definition for Anaconda efiboot.img
Add a small helper function for the efiboot.img partition.  In all
pipelines where we create an Anaconda installer, we use the same single
partition table with the same size, so let's define it in one place.

Note that the variable was incorrectly named `rootfsPartitionTable` in
all cases, but it was in fact the pt for efiboot.img.