test: utility functions for inspecting container images
skopeo_inspect_id() can inspect a container image (local or remote) and
return the image ID for a given architecture.
When inspecting local containers, it uses sudo to access the root
container registry, which is what we use in tests for the bib container.

The host_container_arch() function returns a string that matches the
expected arch name for the host.