CI: test integration to osbuild-composer as part of PRs
It would be useful to explicitly test if the code change proposed in
the PR integrates cleanly to osbuild-composer or if it will require
additional work there. The cheapest way to test it is to try to
replace the osbuild/images ref in the osbuild-composer@main and run
Golang lint and unit tests.

These test cases should not be blocking. The intention is to provide
feedback in the PR, that the author (or someone else) should also work
on integrating the images change to osbuild-composer, since it won't be
"for free".

Signed-off-by: Tomáš Hozza <thozza@redhat.com>