osbuild: fix creating btrfs partitions
This one is a bit weird - the code iterates over all mountable entities, but
the code currently doesn't consider the "root" btrfs (i.e. without
a subvolume) as a mountable entity. Thus, we need to iterate over all entities,
and handle btrfs separately.

A test will be added in a subsequent commit.