osbuild: use New{Group,Users}StageOptions() in Gen{Group,Users}Stage()
We currently duplicate the code that generates the stage options
for users and groups. AFAICT there is no need for this so this
commit just call the existing functions instead of duplicating them.
If we want to remove New{Group,Users}StageOptions() later (which
a previous commit message indicates) we can still do this by just
unexporting or moving the helper.