test: change build info path to include osbuild version
Until now, the following scenario would produce unnecessary rebuilds:
- A PR is opened that updates osbuild in `Schutzfile` to `vNew`.
- A manifest is generated with id `abc123`.
- The build info lookup for `abc123` shows that it was built with `vOld`
  so it's added to the test matrix.
- The new build info is uploaded showing `abc123` was built with `vNew`.
- A different PR runs CI with `vOld` (before the update PR is merged or
  before this PR is rebased).
- A manifest is generated with id `abc123` again.
- The build info lookup for `abc123` shows that it was built with `vNew`
  (in the osbuild update PR) but the currently installed version is
  `vOld` (`!= vNew`) so it's added to the test matrix.
- The new build info is uploaded showing `abc123` was built with `vOld`.

This commit changes the build info paths to include the osbuild package
version (NEVRA) as a path component:

    s3://images-ci-cache/images/builds/fedora-40/aarch64/osbuild-104-1.fc39.noarch/abc123

The build filter now no longer needs to read the osbuild commit ID from
the info.json and compare with the current one.  It can simply check the
path with the package version component and if it is not found, the new
configuration will be tested.