manifest: split stage creation functions for AnacondaInstaller
The stage creation function for AnacondaInstaller (serialize()) was
mostly split between the two types of payload.  Instead of having a
series of switches or condition chains, let's have one and put the stage
creation into separate functions.  There's a tiny bit of duplication for
the dracut stage, but I think that's acceptable for readability.

This commit has no effect on generated manifests.