image: UnattendedKickstart automatically enables ISORootKickstart
The AnacondaTarInstaller image kind supports switching the location of
the kickstart file using the ISORootKickstart option.
When set, the kickstart file is created on the root of the ISO as
/osbuild.ks in the bootiso-tree pipeline.  This is the default for
ostree-based installers and for all installers in the RHEL distros.
When unset, the kickstart file is configured in the interactive defaults
location in the anaconda-tree pipeline (at
/usr/share/anaconda/interactive-defaults.ks).  This is used in Fedora
non-ostree installers.

However, to configure a fully automated (unattended) installation, the
configuration cannot be set in interactive-defaults.ks.  Therefore, if
an unattended installation is needed, if UnattendedKickstart is enabled,
we also automatically enable the ISORootKickstart option.