manifest: common function for handling new kickstart configurations
Move the handling of the UnattendedKickstart and WheelNoPasswd options
to a function to unify handling.  The ostree container case is still
handled separately because its configuration differs slightly.

With this change, the unattended option and the wheel-sudo-nopasswd
options become available to almost all Anaconda-based ISOs.  They are
still not available for the live installer and the ostree container
installer (which is only defined in bootc-image-builder).