image: set kickstart defaults in the images
For the image installers (AnacondaTarInstaller) use the existing OS
customizations which might include distro or image type defaults.  The
effect of this is that the same options are set both in the image and in
the kickstart file.  While this might seem redundant, it enables the
fully unattended installation without changing the configuration in the
payload.