WIP: github: run no-content manifest diff for gitlab trigger
Generate manifests without content to decide if we should trigger builds
in GitLab CI.
- Generate manifests without content in /manifests/HEAD
- Switch to the merge-base with the base branch
- Generate manifests without content in /manifests/merge-base
- Diff the directories

If there are no differences exit with 0, otherwise print a message.

The real version of this job will write a file or set a variable to make
the gitlab trigger skip the gitlab CI if no builds are needed.

Alternatively, it can create the top-level gitlab CI file so that we
don't need to have two nested dynamic pipelines in gitlab.  It can
probably even generate the whole low level one so we don't need to have
dynamic pipelines there and it would make thing a tiny bit saner.

It should also take into account the other things that affect builds:
- osbuild version in Schutzfile
- Seed arg (load it into env when generating manifests)
- bib container ID