osbuild: generate bootupd mounts to span a single disk via partscan=True
The `bootupd` and `bootc install to-filesystem` stages require a
continous disk so that grub can install the bootloader onto the
disk. This means that the new `partscan` feature from [0] must be
used when generating the disk instead of the current approach that
generates a bunch of loopback devices for each partition.

Note that this version of the code does not implement support
for "container" partition tables like LUKS.

[0] https://github.com/osbuild/osbuild/pull/1501