disk: add PartitionTable.GetMountpointSize
It's generally useful to be able to get a size of a certain mountpoint. Let's
add a method for this.

In order to cover this method with tests, I extended the fake PT generator to
correctly set partition sizes. Also, I had to move the current content of
partition_table_test to a separate file and package, in order to break
an inclusion cycle.