distro/rhel8: Adjust /boot size again
It turns out 500 MiB /boot is too small. Thus, we agreed to do the following
changes for 8.10:

- for x86_64, let's not have a separate /boot at all (parity with 8.8 and
  older)
- for aarch64, let's make /boot 1 GiB big

I've removed an old, weird test for EC2 partitioning, and instead implemented
a new one that actually checks whether /boot is correctly sized (or even if it
exists) for a given distro-arch-image type combo.

This also removed different partition table used for the ami image type for
RHEL 8. Before this commit, it always used the "latest" partitioning even for
older minor versions of RHEL 8. Since this created a lot of clutter in the code
(and especially in the newly added test), I've decided to remove this behaviour
and unify it with ec2* images. Thus, an ami image built from an older minor
release will have a different /boot size than the latest minor. I think that's
fine because users can always use a customization to make their /boot bigger.