rhel9/images:Custom service to create mountpoints
Filesystem customization  mountpoints does not persists any upgrade
due to the property of ostree which only persists data on /var and /etc.
This feature will ingest the filesystem customization and create a unit
file which is stored in /etc/systemd/system and can create the required
mountpoints if it does not exist.
fixes: https://github.com/osbuild/images/issues/352

Signed-off-by: Sayan Paul <paul.sayan@gmail.com>