RHEL-8: don't touch rhsm.conf on non-RHUI images.
Previously, the auto-registration was explicitly enabled in rhsm.conf
even for non-RHUI cloud images. These are the images that users can
build on-prem or in the hosted service (as vanilla or as customized).
However, the feature never really worked for these images.

Stop enabling the auto-registration in rhsm.conf (basically stop
touching the configuration at all).

Related to https://issues.redhat.com/browse/COMPOSER-2157

Signed-off-by: Tomáš Hozza <thozza@redhat.com>