blueprint: add a (reverse) path to CustomizationError
Add a reverse path field to the CustomizationError that encodes the
path to the customization that caused an error.

The path is constructed in reverse (bottom-up) for convenience when
iterating or recursing through the blueprint structure and is reversed
to top-down when printing the error message.

This will be used to inform the user of the specific key that was added
to a blueprint but wasn't supported (or required and not set) without
any confusion arising from keys that share a name (e.g. kernel.name and
user.name).