distro: validate supported blueprint options
A validation function that takes a list of supported paths to keys
(which represent blueprint json tags) and a blueprint and checks that
every value set in the blueprint is in the list of supported options.
This validator is similar to the existing Customizations.CheckAllowed()
function but it doesn't only check for values on one level - it can
check for supported values of arbitrary depth.  For example, the
following value:

    customizations.kernel.name

means that the 'name' field of the 'kernel' customization is supported
and if no other item is listed, it also implies that the kernel.append
customization is not supported.

This also works with array types, meaning:

    customizations.user.name
    customizations.user.key

means that the 'name' and 'key' options are supported for each element
of the 'user' array, but none of the other fields are supported.

Error messages use a new ValidationError type to print the path to
the specific option that caused the error.

The new error type contains a reverse path field that encodes the path
to the option that caused an error.

The path is constructed in reverse (bottom-up) for convenience when
iterating or recursing through the structure and is reversed to top-down
when printing the error message.

This will be used to inform the user of the specific key that was added
to a blueprint but wasn't supported (or required and not set) without
any confusion arising from keys that share a name (e.g. kernel.name and
user.name).

This will replace the blueprint.CustomizationError that is currently
used for a similar purpose.