distro: validate supported blueprint options
A validation function that takes a list of supported paths to keys
(which represent blueprint json tags) and a blueprint and checks that
every value set in the blueprint is in the list of supported options.
This validator is similar to the existing Customizations.CheckAllowed()
function but it doesn't only check for values on one level - it can
check for supported values of arbitrary depth.  For example, the
following value:

    customizations.kernel.name

means that the 'name' field of the 'kernel' customization is supported
and if no other item is listed, it also implies that the kernel.append
customization is not supported.

This also works with array types, meaning:

    customizations.user.name
    customizations.user.key

means that the 'name' and 'key' options are supported for each element
of the 'user' array, but none of the other fields are supported.

Error messages use the new BlueprintError type to print the path to
the specific option that caused the error.