distro: validate supported blueprint options
A validation function that takes a list of supported paths to keys and a
blueprint and checks that every value set in the blueprint is in the
list of supported options.  This validator is similar to the existing
Customizations.CheckAllowed() function but it doesn't only check for
values on one level - it can check for supported values of arbitrary
depth.  For example, the following value:

    Customizations.Kernel.Name

means that the Name field of the Kernel customization is supported and
if no other item is listed, it also implies that the Kernel.Append
customization is not supported.

This also works with array types, meaning:

    Customizations.User.Name
    Customizations.User.SSHKey

means that the Name and SSHKey options are supported for the User array,
but none of the other fields are supported.

Error messages use the new CustomizationError type to print the path to
the specific option that caused the error.