distro: validate required blueprint options
A validation function that takes a list of required paths to keys (which
represent json tags) and a blueprint and checks that every value defined
in the required list is set in the blueprint.  This validator operates
in a reverse manner than the validateSupportedConfig() function since it
needs to iterate over the list of required keys and check that they are
set rather than iterate over the keys that are set and check that they
are supported.

The format of the required options is the same as for supported.  For
example, the following value:

    customizations.user.name

means that the 'name' field of the each 'user' defined in the
customizations should be set to a non-empty value.

Non-empty (non-zero) values are only checked for certain types:

    struct, pointer, slice, and string

Other types (int, bool, etc) can have valid user-specified zero values
and for the purposes of the required config validation, are always
considered set.