distro: validate required blueprint options
A validation function that takes a list of required paths to keys and a
blueprint and checks that every value defined in the required list is
set in the blueprint.  This validator operates in a reverse manner than
the validateSupportedConfig() function since it needs to iterate over
the list of required keys and check that they are set rather than
iterate over the keys that are set and check that they are supported.

The format of the required options is the same as for supported.  For
example, the following value:

    Customizations.User.Name

means that the Name field of the each User defined in the Customizations
should be set to a non-empty value.

Non-empty (non-zero) values are only checked for certain types:

    Struct, Pointer, Slice, and String

Other types (Int, Bool, etc) can have valid user-specified zero values.