github: ignore SKIP_CI when triggering GitLab
We never used the SKIP_CI logic in this repository.  We intend to ake
the GitLab workflow a pretty much empty one that's just there to trigger
the chain, but we first need to merge this change so it's in main while
the PR that changes the trigger workflow runs.