Distro: create `rhel` under `distro` for common RHEL
As the first step towards defining el10 and el9 in without completely
forking the `rhel9` distro, create a common `rhel` directory and move
`rhel9` under it. The idea is to then extract the parts that should be
shared across RHEL distros under `rhel` package and export necessary
structs and functions from it and make use of them.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>