Support user-defined entries for sudo NOPASSWD
Replace the wheel-sudo-nopasswd (bool) installer customization with
sudo-nopasswd ([]string).  Users can now specify a list of users or
groups (groups must be prefixed with %).  For each element in the array,
the kickstart file will create a file that enables sudo with NOPASSWD
for that entry.  Each entry is added as a separate file in the sudoers.d
drop-in directory.

Duplicate entries are silently ignored.