image: generate the all-manifest for bootc-disk
The bootc-disk image manifest now always contains all pipelines needed
to generate all image types.  This way we can build any combination of
disk images depending on what we need to export without needing to
conditionally add or remove pipelines.

The Platform no longer needs to be read, since we're generating all of
the formats and the architecture depends on the source container.
However, the Platform is still required for specifying the boot mode
(enabling or disabling BIOS boot).

The first return value of the InstantiateManifestFromContainers()
function, the artifact, was not used and has been removed now.