distro/rhel9: add small subset of image types to EL10
Temporarily separate the image type addition in the rhel9 package to
only add a small subset of image types to EL10 (RHEL 10.0 and CentOS
Stream 10).

For a start, we will only support the following image types on EL10
because of package availability:
- ami
- oci (oracle)
- openstack
- ova
- qcow2
- tar
- vhd
- vmdk
- wsl

This big conditional separation should be removed when most image types
become available in EL10.