test: move the bib ref to the Schutzfile
Move the bib ref into the Schutzfile instead of hard-coding it in the
imgtestlib.  Since bib is distro-agnostic, put it under the `common`
key.

With this change, we'll be able to pin and automatically update the
reference to the bib container so we don't test against moving targets.

The ref used in boot-image no longer needs to come from the local
container storage since we now use a pinned ref.