Distro: add DefaultOSCAPDatastream option to ImageConfig
All RHEL distros used to hardcode the setting of the default
OSCAP Datastream in their distro-specific implementation of
the osCustomizations().

This makes it hard to reuse the same osCustomizations() implementation
across all the RHEL distros.

Therefore, move the DefaultOSCAPDatastream option to ImageConfig, which
will allow to set it in the default distro image configuration. Then the
common osCustomizations() implementation will set it in the
OSCustomizations if the user didn't provide any value in the
Blueprint customizations. In case the user doesn't provide any value and
the distro does not have any default defined, an error is returned.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>