Distro: add KernelOptionsBootloader option to ImageConfig
Old distros (el7 and el8) used to hardcode the setting of
KernelOptionsBootloader option in `OSCustomizations` in their
distro-specific implementation of the osCustomizations().

The option was set for all bootable images on all architectures, except
for the s390x.

This makes it hard to reuse the same osCustomizations() implementation
across all the RHEL distros.

Therefore, move the KernelOptionsBootloader option to ImageConfig, which
will allow to set it in the default distro image configuration. Then the
common osCustomizations() implementation will set it in the
OSCustomizations if the image is bootable. And lastly, the OS pipeline
implementation will handle the architecture check and ignore the option
on s390x.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>