Distro/RHEL-9: introduce ostreeDeploymentCustomizations()
Introduce a helper function ostreeDeploymentCustomizations() and move
the logic, to combine ImageConfig and BP customizations to produces
OSTreeeDeploymentCustomization instance, to it.

Note on the change of the Locale in the raw image ImageConfig - the
value was previously ignored, therefore I set it now to the value that
was previously hardcoded in the image type implementation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>