dnfjson: use repository configs from response to convert packages
Use the repository configurations that osbuild-depsolve-dnf returns to
populate extra repo-related metadata on each package (like CheckGPG).
This will now also include repositories loaded from the RootDir, so
packages are properly configured even when they're retrieved from
repositories not configured in the request.

RHSM is treated as a special property now, outside of other repository
properties and options.  RHSM can only be set on repos defined in the
request, not from repositories in the RootDir, so we keep a set of
repository IDs that specify RHSM in order to set them on each package
that requires it.