Distro: add InstallerConfig data structure
Add a new InstallerConfig data structure, an equivalent to ImageConfig,
but for the installer part of an installer image types.

This will allow to configure some aspects of the installer images on the
image definition level, instead of hardcoding it in the image kind
implementation.

This will make the image kind implementations reusable across multiple
distro versions.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>