many: revert PR#462 to unblock images temporarily
It turns out that we need user customizations that go beyond what
PR#462 provided. So to unblock the other work around the iso installer
revert this PR temporarily and once the groundwork in osbuild is
ready this commit here will be reverted and the pieces added that
enable user customizations with `bootc install to-filesystem`.