many: redo customizations in a `bootc install to-filesystem` world
This commit implements filesystem customizations for images that
got generated via `bootc install to-filesystem` via the new
`org.osbuild.bind` mechanism.

First the image is written via the `bootc.install-to-fs` stage,
then the image is mounted, the ostree.deployment stage is used
to get the right deployment and finally the mount is put over
the "tree" directory via the bind mount module. This way for
the users and selinux stages nothing changes, they work as
before on a normal tree (that just happens to be mounted from an
image this time).

It also includes a stage to make the /var/home directory that is
missing by default on the generic images.