dnfjson: osbuild-depsolve-dnf property rename
Some repo properties were renamed in osbuild-depsolve-dnf to match the
equivalent ones in dnf (libdnf).

Of note is the SSLVerify property which was previously called IgnoreSSL
and had the opposite effect.  The nil value for SSLVerify falls back to
'true' in osbuild-depsolve-dnf which maintains the same behaviour.  We
make the new SSLVerify property a pointer so that we can express the
unset/undefined/nil value properly and let osbuild-depsolve-dnf define
the fallback value.

This change breaks backwards compatibility.