manifest: use new repo configs when serializing manifests
Pass down the repository configurations we get from osbuild-depsolve-dnf
into the manifest serialization functions.  These repositories are used
to add gpg keys to import for each rpm stage in a pipeline.  The
repository configurations are appended to the ones already defined in
each pipeline before serialization.

This will duplicate repositories in each pipeline, effectively
duplicating gpg key importing.  This wont have any effect on the build
or the final image.