osbuild/rpm: generate stable and unique gpg key options
The addition of repo configs from the depsolve result added in PR #537
[1] added a bit of non-determinism to the rpm stage option generation.
The list of gpg keys to import always had duplicates but it was at least
stable, based on the repository configurations for each build.  Now,
the repository configurations that we get from the depsolve aren't in
stable order so the key order can change.  This had no functional effect
on the image build process, but it does mean that manifests generated
with the same inputs have different IDs.

Sort and deduplicate keys in the rpm stage option generation to make
manifests stable.  Deduplicating the keys also makes the manifests a bit
"cleaner".

[1] https://github.com/osbuild/images/pull/537

Signed-off-by: Achilleas Koutsou <achilleas@koutsou.net>