container: include the architecture of the image in container.Spec
This is a different attempt to resolve the issue that the resolver
consider the requested architecture more of a "hint". The initial
attempt is in https://github.com/osbuild/images/pull/451 but
it caused some issues as this seems to be expected behavior of
the library in tests at least.

So this commit is more conservative and just includes the the
architecture in the `container.Spec` so that the consumer of
the library can inspect and error on mismatch of requested
arch and resolved arch.