manifest: write custom `/etc/fstab` in `RawBootcImage`
With the new uniform way to handle writing to the bootc image deployment
we can now support a custom `/etc/fstab` again. Similar to what we
do for users [0] and groups [1] we allow also writing a custom fstab
now.

This will need https://github.com/osbuild/osbuild/pull/1727 and also
probably a port of the fstab module to schema_2.

[0] https://github.com/osbuild/images/pull/571
[1] https://github.com/osbuild/images/pull/593