Do not ignore "package_sets" from repo configs
Previously, the "package_sets" property of repo configs was ignored when
the config was read from the file. The property is used at least by
RHEL-7 images, which install additional packages from the eng-rel
repository when preparing the buildroot container. Previously, the
eng-rel repository would be used also for the image base package set and
would taint the resulting manifest and image.

Modify the function loading repo configs from a file to not ignore the
"package_sets" property. Fix the property name in "rhel-7.9.json" config.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>