image: bring back the pre-PR#462 `BootcDiskImage`
This commit brings back the pre-PR#462 version of BootcDiskImage as
BootcLegacyDiskImage. The rational is that right now we cannot
use `bootc install to-filesystem` for cross arch building the
FITRIM ioctl is not supported by qemu-user

To fix this is requires any of the following:
1. ensure https://github.com/containers/bootc/pull/462 is available
   everywhere we care about
2. ensure https://www.mail-archive.com/qemu-devel@nongnu.org/msg1034508.html
   is available everywhere we care about
3. bind mount /usr/bin/true over /usr/bin/fsfreeze inside the buildroot.
   This bind mount can be done via `org.osbuild.bind` but it leaks into
   the manifest and that does not feel great.

So the alternative is to keep using the "old" pipeline that does not
yet use `bootc install to-filesystem` to for cross-arch builds and
the new pipeline for everything else. The selection will happen
on the level of `bib`.

This is all not nice and fragile and needs to be removed/reverted
again but short term might be the only thing we can do.