stage-systemd-unit-create:fix defaultdepencies flag Added a function to pass the bool as pointer.
DefaultDependencies in systemd  is by default set as true and needs to be
exluscively set as false which is oppositeof  golangs syntax omitempty.
Leaving out omitempty causes the default dependencies to be set as
false by default which will cause the service to behave abnormally as its
by default set to true in systemd world.

Signed-off-by: Sayan Paul <paul.sayan@gmail.com>