Distro/el7/ec2: fix kernel cmdline
The kernel cmdline was missing the `ro`, `crashkernel=auto` and
`LANG=en_US.UTF-8`, which were not explicitly specified in the KS.

Moreover, it turns out that `image-info` looks for bootloader entries in
`/boot/loader/entries/*`, but that path does not exist on RHEL-7.

Modify the kernel cmdline used for the ec2 image based on the content of
`/boot/grub2/grub.cfg` of the original RHEL-7.9 image built using
ImageFactory.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>