manifest: inject env file with org and activation key
Create the file in /etc that will act as both the EnvironmentFile and
the ConditionPathExists for the service.  This file includes the org ID
and activation key.  When the service runs, the file is removed so that
the service doesn't run again (only runs on first boot) and the org ID
and key are removed from the system.

Rename the file to osbuild-subscription-register.env to make its purpose
clearer.