container: fix crash when `Container.New()` defer() is run
This commit fixes a nil pointer exception when `Container.New()`
needs to run the defer that calls c.Stop(). The semantics of named
returns are subtle here. The local var `c` is non-nil. However
when `New()` returns with `return nil, err` (e.g. in line 64+)
then `c` becomes `nil` (because the of return nil). Which means
the defer now tries to use a `nil` pointer and crashes.

The fix is to either use `naked` returns everywhere or just
remove the named returns. The later seems more in line with what
we (and the go community) generally do.

A regression test is added too.