distro: wire up the user kickstart customization
Pass the user kickstart customization from the blueprint options into
the pipeline generators and append it to any hardcoded kickstart bits
that we might be including ourselves.  Just like with other injected
kickstart parts, like the %post section we add for sudoers, the user's
kickstart file, when specified, will be added to an iso as osbuild.ks
and the kickstart file created by osbuild will be renamed to
osbuilld-base.ks and run using an %include added to the top of the
osbuild.ks file.