pkg/container: tidy up logic for override digest
The `cl.GetManifest` function is called recursively, if it is a simple
image the image digest will be empty. If it is called with a manifest
list, the `instanceDigest` will not be empty. We can tidy this function
up a little and make the logic a little clearer and easier to
understand.

Additionally, since local manifest lists are handled differently to
manifest lists in remote repositories, we need to handle this case too
and not set the `overrideDigest`.

Co-authored-by: Michael Vogt <michael.vogt@gmail.com>