gen-manifests: derrive seedArg from filename and env
To ensure that manifests get random(ish) and stable UUIDs we set
the rng seed arg based on the filename.

This should fix the issue discovered in
osbuild/manifest-db#124

that duplicated UUIDs for xfs/btrfs can trigger random(ish) and
hard to diagnose errors.

This is the same as
https://github.com/osbuild/osbuild-composer/pull/4124
hopefully for the right place this time.