test/scripts: fail setup-osbuild-repo on bad repo
When testing with "pinned" osbuild commit IDs (defined in the
Schutzfile), we set up a repository based on the commit ID that should
exist for each commit in osbuild that had a (partially) successful CI
run (the RPM jobs should at least run successfully).

When a repository doesn't exist for whatever reason (the CI pipeline
failed, the commit ID was copied incorrectly, etc), the setup script
wouldn't fail and osbuild would be installed from the system
repositories.  This means that we're not testing what we think we are
testing.  In the best case our build jobs fail because the osbuild
version in the repos isn't new enough to have all the features we're
testing, but we might also end up having passing tests that aren't
testing what we want, for example a change in osbuild internals that's
on an open PR and we want to verify that none or our test image builds
are broken.

Make the setup-osbuild-repo script check that the repository metadata is
reachable and exit with an error if it's not, making any test pipeline
that requires it fail when necessary instead of silently skipping the
repo configuration.