policies: deny more paths for ostree-based images
Adding more paths that are denied for mountpoint creation for
ostree-based images.  In a booted ostree system, these paths are
symlinks, so it makes no sense (and will cause problems) to create
mountpoints at those locations.

The list has been sorted to make it simpler to modify consistently going
forward.